Rename test files to be unique
